Revert "Workerized emscripten retroarch (WIP) (#17484)"
This reverts commit cacd5a9a234f3c9abc8d352f6fea19117e2e1a00.
