+++ Makefile.common
@@ -1526,10 +1526,7 @@ ifeq ($(HAVE_GL_CONTEXT), 1)
    endif
 
    ifeq ($(HAVE_EMSCRIPTEN), 1)
-      ifeq ($(HAVE_EGL), 1)
-        OBJ += gfx/drivers_context/emscriptenegl_ctx.o
-      endif
-      OBJ += gfx/drivers_context/emscriptenwebgl_ctx.o
+      OBJ += gfx/drivers_context/emscriptenegl_ctx.o
    endif
 
    ifeq ($(HAVE_MALI_FBDEV), 1)
