+++ gfx/drivers_context/emscriptenegl_ctx.c
@@ -124,9 +124,11 @@ static void gfx_ctx_emscripten_destroy(void *data)
 
    if (!emscripten)
       return;
+
 #ifdef HAVE_EGL
    egl_destroy(&emscripten->egl);
 #endif
+
    free(data);
 }
 
@@ -189,6 +191,7 @@ static void *gfx_ctx_emscripten_init(void *video_driver)
 #endif
 
    return emscripten;
+
 error:
    gfx_ctx_emscripten_destroy(video_driver);
    return NULL;
