+++ libretro-common/include/retro_timers.h
@@ -39,7 +39,7 @@
 #include <psp2/kernel/threadmgr.h>
 #elif defined(_3DS)
 #include <3ds.h>
-#elif (defined(EMSCRIPTEN) && defined(EMSCRIPTEN_ASYNCIFY))
+#elif defined(EMSCRIPTEN)
 #include <emscripten/emscripten.h>
 #else
 #include <time.h>
@@ -100,7 +100,7 @@ static int nanosleepDOS(const struct timespec *rqtp, struct timespec *rmtp)
 #define retro_sleep(msec) (usleep(1000 * (msec)))
 #elif defined(WIIU)
 #define retro_sleep(msec) (OSSleepTicks(ms_to_ticks((msec))))
-#elif defined(EMSCRIPTEN) && defined(EMSCRIPTEN_ASYNCIFY)
+#elif defined(EMSCRIPTEN)
 #define retro_sleep(msec) (emscripten_sleep(msec))
 #else
 static INLINE void retro_sleep(unsigned msec)
