+++ pkg/emscripten/libretro/index.html
@@ -192,13 +192,13 @@
             <img class="webplayer-preview img-fluid" src="media/canvas.png" width="960" height="720" alt="RetroArch Logo">
          </div>
       </div>
-      <script crossorigin="anonymous" src="//code.jquery.com/jquery-3.1.0.min.js"></script>
-      <script crossorigin="anonymous" src="//rawgit.com/jeresig/jquery.hotkeys/master/jquery.hotkeys.js"></script>
-      <script crossorigin="anonymous" src="//cdnjs.cloudflare.com/ajax/libs/tether/1.3.4/js/tether.min.js"></script>
-      <script crossorigin="anonymous" src="//cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.0.0-alpha.3/js/bootstrap.min.js"></script>
+      <script src="//code.jquery.com/jquery-3.1.0.min.js"></script>
+      <script src="//rawgit.com/jeresig/jquery.hotkeys/master/jquery.hotkeys.js"></script>
+      <script src="//cdnjs.cloudflare.com/ajax/libs/tether/1.3.4/js/tether.min.js"></script>
+      <script src="//cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.0.0-alpha.3/js/bootstrap.min.js"></script>
       <script src="analytics.js"></script>
       <!--script src="//wzrd.in/standalone/browserfs@0.6.1"></script-->
-      <script type="text/javascript" src="zip.min.js"></script>
+      <script src="browserfs.min.js"></script>
       <script src="libretro.js"></script>
    </body>
 </html>
